requirements.txt
tensorflow>=2.13.0
keras>=2.13.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=1.5.3
scikit-learn>=1.3.0
pillow>=10.0.0
streamlit>=1.30.0
joblib>=1.3.0
altair>=5.0.0